added a console runner with some output, playing around with learning rate

+++ NeuralNetworkCSharp/ActivationFunctions/StepActivationFunction.cs
@@ -22,7 +22,7 @@ namespace NeuralNetworkCSharp.ActivationFunctions
 
         public double CalculateOutput(double input)
         {
-            return Convert.ToDouble(input > _treshold);
+            return Convert.ToSingle(input > _treshold);
         }
     }
 }
